requirements.txt
ultralytics>=8.0.0
coremltools>=7.0
numpy>=1.24.0
Pillow>=10.0.0 